Normm(parser): subdivided the functions in different files
